# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Press Kits Service",
    "description": "API for generating, versioning, and publicly sharing company press kits",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://press-kits.mcpfactory.org"
    }
  ],
  "components": {
    "schemas": {
...
      "ViewStatsFlatResponse": {
        "type": "object",
        "properties": {
          "totalViews": {
            "type": "number",
            "example": 1250
          },
          "uniqueVisitors": {
            "type": "number",
            "example": 843
          },
          "lastViewedAt": {
            "type": "string",
            "nullable": true,
            "example": "2026-03-29T14:32:00.000Z"
          },
          "firstViewedAt": {
            "type": "string",
            "nullable": true,
            "example": "2026-03-01T09:15:00.000Z"
          }
        },
        "required": [
          "totalViews",
          "uniqueVisitors",
          "lastViewedAt",
          "firstViewedAt"
        ]
      },
      "ViewStatsGroupedResponse": {
        "type": "object",
        "properties": {
          "groups": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "key": {
                  "type": "string",
                  "nullable": true,
                  "description": "Group key (country code, kit ID, or date depending on groupBy)",
                  "example": "US"
                },
                "totalViews": {
                  "type": "number",
                  "example": 520
                },
                "uniqueVisitors": {
                  "type": "number",
                  "example": 340
                },
                "lastViewedAt": {
                  "type": "string",
                  "nullable": true,
                  "example": "2026-03-29T14:32:00.000Z"
                }
              },
              "required": [
                "key",
                "totalViews",
                "uniqueVisitors",
                "lastViewedAt"
              ]
            }
          }
        },
        "required": [
          "groups"
        ]
      },
      "ViewStatsGroupBy": {
        "type": "string",
        "enum": [
          "country",
          "mediaKitId",
          "day",
          "brandId",
          "campaignId",
          "featureSlug",
          "workflowSlug",
          "featureDynastySlug",
          "workflowDynastySlug"
        ],
        "description": "Group results by dimension. Omit for flat totals."
      },
      "CostStatsGroup": {
        "type": "object",
        "properties": {
          "dimensions": {
            "type": "object",
            "additionalProperties": {
              "type": "string",
              "nullable": true
            },
            "description": "Group key dimensions"
          },
          "totalCostInUsdCents": {
            "type": "number",
            "example": 2050
          },
          "actualCostInUsdCents": {
            "type": "number",
            "example": 2050
          },
          "provisionedCostInUsdCents": {
            "type": "number",
            "example": 0
          },
          "runCount": {
            "type": "number",
            "example": 3
          }
        },
        "required": [
          "dimensions",
          "totalCostInUsdCents",
          "actualCostInUsdCents",
          "provisionedCostInUsdCents",
          "runCount"
        ]
      },
      "CostStatsResponse": {
        "type": "object",
        "properties": {
          "groups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CostStatsGroup"
            }
          }
        },
        "required": [
          "groups"
        ]
      },
      "CostStatsGroupBy": {
        "type": "string",
        "enum": [
          "mediaKitId",
          "brandId",
          "campaignId",
          "featureSlug",
          "workflowSlug",
          "featureDynastySlug",
          "workflowDynastySlug"
        ],
        "description": "Group results by dimension. Omit for flat totals."
      },
...
    },
...
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
  "paths": {
...
    "/media-kits/stats/views": {
      "get": {
        "summary": "Get view stats for media kits",
        "description": "Returns aggregated view stats for the org's media kits. Supports filters (brandId, campaignId, mediaKitId, date range) and optional groupBy (country, mediaKitId, day). Without groupBy returns flat totals; with groupBy returns grouped results.",
        "tags": [
          "Stats"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Filter by brand UUID"
            },
            "required": false,
            "name": "brandId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Filter by campaign UUID"
            },
            "required": false,
            "name": "campaignId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Filter by specific media kit UUID"
            },
            "required": false,
            "name": "mediaKitId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Filter by feature slug (e.g. 'press-kit-v2')"
            },
            "required": false,
            "name": "featureSlug",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Filter by workflow slug"
            },
            "required": false,
            "name": "workflowSlug",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Filter by feature dynasty slug (stable across versions)"
            },
            "required": false,
            "name": "featureDynastySlug",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Filter by workflow dynasty slug (stable across versions)"
            },
            "required": false,
            "name": "workflowDynastySlug",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "date-time",
              "description": "Start of date range (ISO 8601)",
              "example": "2026-03-01T00:00:00Z"
            },
            "required": false,
            "name": "from",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "date-time",
              "description": "End of date range (ISO 8601)",
              "example": "2026-03-31T23:59:59Z"
            },
            "required": false,
            "name": "to",
            "in": "query"
          },
          {
            "schema": {
              "$ref": "#/components/schemas/ViewStatsGroupBy"
            },
            "required": false,
            "name": "groupBy",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Internal org UUID from client-service",
              "example": "org_3ANNRtJtvq2vahygqOSJ7IjRfp1"
            },
            "required": true,
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Internal user UUID from client-service",
              "example": "usr_7BXkm2pTq1wLnHyjRfK4d"
            },
            "required": true,
            "name": "x-user-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Run ID for tracing (from runs-service)",
              "example": "550e8400-e29b-41d4-a716-446655440000"
            },
            "required": true,
            "name": "x-run-id",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "View stats (flat or grouped depending on groupBy param)",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ViewStatsFlatResponse"
                    },
                    {
                      "$ref": "#/components/schemas/ViewStatsGroupedResponse"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/media-kits/stats/costs": {
      "get": {
        "summary": "Get cost stats for media kit generations",
        "description": "Returns aggregated generation/edit costs by querying runs-service for all runs associated with media kits. Supports filters (mediaKitId, brandId, campaignId) and optional groupBy (mediaKitId). Without groupBy returns flat totals; with groupBy returns grouped results.",
        "tags": [
          "Stats"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Filter by specific media kit UUID"
            },
            "required": false,
            "name": "mediaKitId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Filter by brand UUID"
            },
            "required": false,
            "name": "brandId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Filter by campaign UUID"
            },
            "required": false,
            "name": "campaignId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Filter by feature slug (e.g. 'press-kit-v2')"
            },
            "required": false,
            "name": "featureSlug",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Filter by workflow slug"
            },
            "required": false,
            "name": "workflowSlug",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Filter by feature dynasty slug (stable across versions)"
            },
            "required": false,
            "name": "featureDynastySlug",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Filter by workflow dynasty slug (stable across versions)"
            },
            "required": false,
            "name": "workflowDynastySlug",
            "in": "query"
          },
          {
            "schema": {
              "$ref": "#/components/schemas/CostStatsGroupBy"
            },
            "required": false,
            "name": "groupBy",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Internal org UUID from client-service",
              "example": "org_3ANNRtJtvq2vahygqOSJ7IjRfp1"
            },
            "required": true,
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Internal user UUID from client-service",
              "example": "usr_7BXkm2pTq1wLnHyjRfK4d"
            },
            "required": true,
            "name": "x-user-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Run ID for tracing (from runs-service)",
              "example": "550e8400-e29b-41d4-a716-446655440000"
            },
            "required": true,
            "name": "x-run-id",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Cost stats (flat or grouped depending on groupBy param)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CostStatsResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "security": [
    {
      "apiKey": []
    }
  ]
}